fix: don't delete fresh folders

+++ src/scrubber.rs
@@ -5,6 +5,7 @@
 use std::io::ErrorKind;
 use std::path::PathBuf;
 use std::sync::Arc;
+use std::time::{Duration, SystemTime};
 
 use chrono::Utc;
 use log::{error, info, warn};
@@ -103,8 +104,18 @@ fn _cleanup_folder(path: &PathBuf) -> anyhow::Result<()> {
 
     let is_empty = path.read_dir()?.next().is_none();
     if is_empty {
-        info!("Scrubber: {:?}: Deleting empty folder", path);
-        std::fs::remove_dir(path)?;
+        let created = path.metadata()?.created()?;
+        let now = SystemTime::now();
+
+        // This can Err if the clock goes backwards.
+        // We ignore that and hope that eventually in a later scrub it won't go back
+        // At least not enough to trigger an error
+        if let Ok(duration) = now.duration_since(created) {
+            if duration > Duration::from_secs(60 * 60) {
+                info!("Scrubber: {:?}: Deleting empty folder", path);
+                std::fs::remove_dir(path)?;
+            }
+        }
     }
 
     Ok(())
